docs/AI编程实战/ai-code-reviewer/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ai-code-reviewer"
version = "1.0.0"
description = "AI-powered code review tool supporting multiple analysis dimensions"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "rich>=13.0.0",
    "gitpython>=3.1.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
ai-review = "ai_code_reviewer.cli:main"

[tool.setuptools.packages.find]
include = ["ai_code_reviewer*"]
